Folder structure rebuild with CLI's
